refactor: extract formula LaTeX display logic to separate utility function
